project-root/requirements.txt
python-dotenv>=1.0.0
requests>=2.28
python-telegram-bot>=20.0
pydantic>=1.10
aiohttp>=3.8
openai>=0.27.0    # opcional: só se for usar LLM
websockets>=10.4  # opcional para conexões RPC/streams
